app/requirements.txt
sqlmodel==0.0.24
typer==0.15.2
ipython==9.0.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
requests==2.32.3
ollama==0.4.7